+++ internal/web/handlers/url_test.go
@@ -11,8 +11,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
 	"github.com/shulganew/shear.git/internal/config"
-	"github.com/shulganew/shear.git/internal/internal/mocks"
-	"github.com/shulganew/shear.git/internal/service"
+	"github.com/shulganew/shear.git/internal/service/mocks"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 )
@@ -72,10 +71,8 @@ func TestURL(t *testing.T) {
 			configApp.Address = config.DefaultHost
 			configApp.Response = config.DefaultHost
 
-			stor := service.StorageURL(storeMock)
-
 			//init storage
-			handler := NewHandlerWeb(configApp, &stor)
+			handler := NewHandlerWeb(configApp, storeMock)
 
 			userID, err := uuid.NewV7()
 			if err != nil {
