+++ internal/web/middlewares/middlewgzip.go
